# perf: remove not_intersected_yet