This includes #84, the only new part is the last commit. It is intended to be merged after it.

For a Range an intersection is almost as fast as a clone. By removing the not_intersected_yet optimization we do more intersections and less clones and allocations. This simplifies the code for `term_intersection_for_package` witch is in the hotest part of the of `unit_propagation`.

I was trying this because of how much it simplified the code. It would be worth doing even if it was perf neutral, but I am consistently seeing small perf improvements!

```
Benchmarking large_cases/elm-packages_str_SemanticVersion.ron: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 20.0s. You may wish to increase target time to 28.2s, or reduce sample count to 70.
large_cases/elm-packages_str_SemanticVersion.ron
                        time:   [279.06 ms 279.37 ms 279.67 ms]
                        change: [-6.6904% -6.5466% -6.4031%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
large_cases/large_case_u16_NumberVersion.ron
                        time:   [10.604 ms 10.634 ms 10.670 ms]
                        change: [-10.819% -10.537% -10.273%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high severe
Benchmarking large_cases/zuse_str_SemanticVersion.ron: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 20.0s. You may wish to increase target time to 38.2s, or reduce sample count to 50.
large_cases/zuse_str_SemanticVersion.ron
                        time:   [380.02 ms 380.46 ms 380.92 ms]
                        change: [-8.5621% -8.3803% -8.2179%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild
  2 (2.00%) high severe
```