@mpizenberg 2021-05-01
So this seems like a follow up to #33 that was done quite a while ago.

I feel like this could go even further and simplify some of the parameters around the `relation` function. Currently, we have `pub fn relation(&self, terms: impl Fn(&P) -> Option<&'a Term<V>>)`. If I remember correctly, the name `terms` comes from the fact that it used to be a function returning an iterator or a slice of terms. Then we decided to precompute terms intersection, but only when necessary, while the rest of the time terms are accumulated in a vector `not_intersected_yet`, awaiting to be useful. And when we decided that, we only had the resulting intersection to pass to the function, thus `Option<Term<V>>`. But now that intersection are systematically computed, I feel like this could be turned into `Fn(&P) -> &Term<V>` without the option. And when there is no term, the intersection is anything (was it `Term::any()?` I forgot). The point is we could have that defined as a static reference to a constant term.

All that is just gut feeling based on what I remember of all this. I think you have all this in mind better than me now so I hope you'll see what I'm trying to say ^^.