@Eh2406 2021-05-03
You are correct. `Option<&'a Term<V>>` can be replaced with `&'a Term<V>` by `unwrap_or(&Term::any())`. ( Borrow checker makes thinks a little more complicated, but it is not relevant. ) And I was about to push it, but let me be thorough and run new perf numbers ... and that is not good. By returning `None` we skip calling `relation_with` witch is important for performance.