@mpizenberg 2021-05-03
You're right! I forgot about the cost of that call. Seems great to me then!